% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rempalindromesnp.R
\name{rmpalsnp}
\alias{rmpalsnp}
\title{Remove Palindromic SNPs}
\usage{
rmpalsnp(testData, t)
}
\arguments{
\item{testData}{data.frame or data.table containing "reference_allele" and "other_allele" columns, each column contains single character (either A, T, G, or C). Note that the function uses rs_number to remove the SNPs from the original data.frame, making "rs_number" a required column. Allele frequency column should be named "eaf".}

\item{t}{Threshold value >0 & <=0.5}
}
\value{
data.table or data.frame with removed palindromic SNPs within 0.5 +/- threshold t (note that function must be called and assigned to a data.frame, see examples)
}
\description{
Detect SNPs palindromic with palindromic alleles and remove them if their allele frequency is within a specific threshold around the allele frequency of 0.5.
}
\examples{
result_df <- rmpalsnp(input_df, t);
result_df <- rmpalsnp(myData, 0.2);
}
